Everyone is sleep but work must continue

+++ model/creeper-simulation-library/Field.cpp
@@ -44,7 +44,7 @@ Field::Field(Point size, double r0, size_t creepersNum, double moveRadius,
 
 void Field::updateField() {
 #pragma omp parallel for
-  for (size_t i = 0; i < creepers_.size(); ++i) {
+  for (int i = 0; i < creepers_.size(); ++i) {
     creepers_[i].walk(generatePosition_);
   }
   auto dist = std::uniform_int_distribution<size_t>(0, creepers_.size());
@@ -59,7 +59,7 @@ void Field::updateField() {
   const double radius = r_0_ * r_0_;
 
 #pragma omp parallel for
-  for (size_t idx = 0; idx < creepersChangingState.size(); ++idx) {
+  for (int idx = 0; idx < creepersChangingState.size(); ++idx) {
     auto i = creepersChangingState[idx];
     for (const auto& creeper2 : creepers_) {
       if (&creepers_[i] != &creeper2) {
